requirements.txt
pyserial
vk_api
todoist-python
